requirements.txt
tensorflow
opencv-python
joblib
sklearn
Pillow

